refactor loop into smaller files
